// src/actions/create-user/schema.ts
import { z } from "zod";

export const createUserSchema = z.object({
  name: z.string().min(3, "O nome deve ter pelo menos 3 caracteres"),
  email: z.string().email("Email inválido"),
  password: z.string().min(6, "A senha deve ter pelo menos 6 caracteres"),
  isActive: z.boolean().default(true),
  
});